+++ src/coq_elpi_vernacular.mli
@@ -69,26 +69,20 @@ type ('a,'b,'c,'d,'e,'f) arg =
   | Qualid of qualified_name
   | DashQualid of qualified_name
   | Term of 'a
-  | EConstr of EConstr.t list
-  | LtacString of 'f
-  | LtacInt of 'f
-  | LtacTermList of 'f
+  | LTac of Coq_elpi_arg_HOAS.ltac_ty * 'f
   | RecordDecl of 'b
   | IndtDecl of 'c
   | ConstantDecl of 'd
   | Context of 'e
 
 type raw_arg = (Constrexpr.constr_expr,  expr_record_decl, expr_indt_decl, expr_constant_decl,Constrexpr.local_binder_expr list,Constrexpr.constr_expr) arg
-type glob_arg = (Genintern.glob_constr_and_expr, Coq_elpi_goal_HOAS.glob_record_decl, Coq_elpi_goal_HOAS.glob_indt_decl, Coq_elpi_goal_HOAS.glob_constant_decl,Coq_elpi_goal_HOAS.glob_context_decl,Glob_term.glob_constr) arg
-type parsed_arg = (Coq_elpi_goal_HOAS.parsed_term, Coq_elpi_goal_HOAS.parsed_record_decl, Coq_elpi_goal_HOAS.parsed_indt_decl, Coq_elpi_goal_HOAS.parsed_constant_decl, Coq_elpi_goal_HOAS.parsed_context_decl, Coq_elpi_goal_HOAS.parsed_term) arg
+type glob_arg = (Genintern.glob_constr_and_expr, Coq_elpi_arg_HOAS.glob_record_decl, Coq_elpi_arg_HOAS.glob_indt_decl, Coq_elpi_arg_HOAS.glob_constant_decl,Coq_elpi_arg_HOAS.glob_context_decl,Glob_term.glob_constr) arg
+type parsed_arg = (Coq_elpi_arg_HOAS.parsed_term, Coq_elpi_arg_HOAS.parsed_record_decl, Coq_elpi_arg_HOAS.parsed_indt_decl, Coq_elpi_arg_HOAS.parsed_constant_decl, Coq_elpi_arg_HOAS.parsed_context_decl, Geninterp.interp_sign * Glob_term.glob_constr) arg
 
 val pr_arg : ('a -> Pp.t) -> ('b -> Pp.t) -> ('c -> Pp.t) -> ('d -> Pp.t) -> ('e -> Pp.t) -> ('f -> Pp.t) -> ('a,'b,'c,'d,'e,'f) arg -> Pp.t
 val glob_arg : Genintern.glob_sign -> raw_arg -> glob_arg
-val interp_arg : Geninterp.interp_sign -> 'g Evd.sigma -> ('a,'b,'c,'d,'e,Glob_term.glob_constr) arg -> Evd.evar_map * (Geninterp.interp_sign * 'a, Geninterp.interp_sign * 'b, Geninterp.interp_sign * 'c, Geninterp.interp_sign * 'd, Geninterp.interp_sign * 'e, 'f1) arg
-val subst_record_decl : Mod_subst.substitution -> Coq_elpi_goal_HOAS.glob_record_decl -> Coq_elpi_goal_HOAS.glob_record_decl
-val subst_indt_decl : Mod_subst.substitution -> Coq_elpi_goal_HOAS.glob_indt_decl -> Coq_elpi_goal_HOAS.glob_indt_decl
-val subst_constant_decl : Mod_subst.substitution -> Coq_elpi_goal_HOAS.glob_constant_decl -> Coq_elpi_goal_HOAS.glob_constant_decl
-val subst_context_decl : Mod_subst.substitution -> Coq_elpi_goal_HOAS.glob_context_decl -> Coq_elpi_goal_HOAS.glob_context_decl
+val interp_arg : Geninterp.interp_sign -> 'g Evd.sigma -> ('a,'b,'c,'d,'e,'f) arg -> Evd.evar_map * (Geninterp.interp_sign * 'a, Geninterp.interp_sign * 'b, Geninterp.interp_sign * 'c, Geninterp.interp_sign * 'd, Geninterp.interp_sign * 'e, Geninterp.interp_sign * 'f) arg
+val subst_arg : Mod_subst.substitution -> glob_arg -> glob_arg
 
 val run_program : Loc.t -> qualified_name -> atts:Attributes.vernac_flags -> raw_arg list -> unit
 val run_in_program : ?program:qualified_name -> Elpi.API.Ast.Loc.t * string -> unit